Refactor/update migration

See merge request personal/projects/moinex!4

+++ src/main/java/org/moinex/model/investment/TickerSale.java
@@ -33,7 +33,7 @@ public class TickerSale extends Transaction {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     @Setter(AccessLevel.NONE)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "ticker_id", referencedColumnName = "id")
@@ -46,7 +46,7 @@ public class TickerSale extends Transaction {
      * Constructor for testing purposes
      */
     public TickerSale(
-            Long id,
+            Integer id,
             Ticker ticker,
             BigDecimal quantity,
             BigDecimal unitPrice,
