+++ src/main/java/org/moinex/repository/investment/CryptoExchangeRepository.java
@@ -11,4 +11,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CryptoExchangeRepository extends JpaRepository<CryptoExchange, Long> {}
+public interface CryptoExchangeRepository extends JpaRepository<CryptoExchange, Integer> {}
