+++ src/main/java/org/moinex/repository/investment/TickerPurchaseRepository.java
@@ -11,4 +11,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TickerPurchaseRepository extends JpaRepository<TickerPurchase, Long> {}
+public interface TickerPurchaseRepository extends JpaRepository<TickerPurchase, Integer> {}
