+++ src/main/java/org/moinex/ui/dialog/wallettransaction/RemoveTransactionController.java
@@ -181,7 +181,7 @@ public class RemoveTransactionController {
      * Configures the TableView to display the incomes.
      */
     private void configureTableView() {
-        TableColumn<WalletTransaction, Long> idColumn = getWalletTransactionLongTableColumn();
+        TableColumn<WalletTransaction, Integer> idColumn = getWalletTransactionLongTableColumn();
 
         TableColumn<WalletTransaction, String> categoryColumn = new TableColumn<>("Category");
         categoryColumn.setCellValueFactory(
@@ -222,8 +222,8 @@ public class RemoveTransactionController {
         transactionsTableView.getColumns().add(statusColumn);
     }
 
-    private static TableColumn<WalletTransaction, Long> getWalletTransactionLongTableColumn() {
-        TableColumn<WalletTransaction, Long> idColumn = new TableColumn<>("ID");
+    private static TableColumn<WalletTransaction, Integer> getWalletTransactionLongTableColumn() {
+        TableColumn<WalletTransaction, Integer> idColumn = new TableColumn<>("ID");
         idColumn.setCellValueFactory(param -> new SimpleObjectProperty<>(param.getValue().getId()));
 
         // Align the ID column to the center
@@ -231,7 +231,7 @@ public class RemoveTransactionController {
                 column ->
                         new TableCell<>() {
                             @Override
-                            protected void updateItem(Long item, boolean empty) {
+                            protected void updateItem(Integer item, boolean empty) {
                                 super.updateItem(item, empty);
                                 if (item == null || empty) {
                                     setText(null);
