+++ src/main/java/org/moinex/ui/main/SavingsController.java
@@ -668,7 +668,7 @@ public class SavingsController {
      * Configure the table view columns
      */
     private void configureTableView() {
-        TableColumn<Ticker, Long> idColumn = getTickerLongTableColumn();
+        TableColumn<Ticker, Integer> idColumn = getTickerLongTableColumn();
 
         TableColumn<Ticker, String> nameColumn = new TableColumn<>("Name");
         nameColumn.setCellValueFactory(
@@ -721,8 +721,8 @@ public class SavingsController {
         stocksFundsTabTickerTable.getColumns().add(avgUnitColumn);
     }
 
-    private static TableColumn<Ticker, Long> getTickerLongTableColumn() {
-        TableColumn<Ticker, Long> idColumn = new TableColumn<>("ID");
+    private static TableColumn<Ticker, Integer> getTickerLongTableColumn() {
+        TableColumn<Ticker, Integer> idColumn = new TableColumn<>("ID");
         idColumn.setCellValueFactory(param -> new SimpleObjectProperty<>(param.getValue().getId()));
 
         // Align the ID column to the center
@@ -730,7 +730,7 @@ public class SavingsController {
                 column ->
                         new TableCell<>() {
                             @Override
-                            protected void updateItem(Long item, boolean empty) {
+                            protected void updateItem(Integer item, boolean empty) {
                                 super.updateItem(item, empty);
                                 if (item == null || empty) {
                                     setText(null);
